refactoring: 폴더 구조 변경 완료, 테스트 완료
