+++ src/context/server.ts
@@ -12,7 +12,7 @@ import { expressMiddleware } from '@apollo/server/express4';
 import cors from 'cors';
 import dotenv from 'dotenv';
 import { GraphQLFormattedError } from "graphql";
-import { createLoaders } from './graphql/context/context';
+import { createLoaders } from './context';
 
 dotenv.config();
