--- src/graphql/context/context.ts
@@ -1,7 +1,0 @@
-import { orderItemsLoader } from '../loaders/orderItemsLoader';
-import { ordersLoader } from '../loaders/ordersLoader';
-
-export const createLoaders = () => ({
-    ordersLoader: ordersLoader(),
-    orderItemsLoader: orderItemsLoader(),
-});
